change type

+++ BalancoTrimestral.java
@@ -6,7 +6,7 @@ class BalancoTrimestral {
 
         int gastosTrimestre = gastosJaneiro + gastosFevereiro + gastosMarco;
 
-        double mediaMensal = gastosTrimestre / 3;
+        float mediaMensal = gastosTrimestre / 3;
 
         System.out.println("Gastos trimestre: " + gastosTrimestre);
         System.out.println("Valor da média mensal: " + mediaMensal);
